samples/ds-ml-ai--data-science-machine-learning-artificial-intelligence/11-etl-extract-transform-load/extractions/images-from-documents/pdf/minecart/requirements.txt
codetiming==1.4.0
minecart==0.3.0
orjson==3.11.7
pdfminer3k==1.3.4
pillow==12.1.1
ply==3.11
six==1.17.0
timer==0.3.0
